修改了原有的网络参数，本版本可顺利跑通，目前三组准确率在40%左右，需要调参

+++ Utils/ReadAndDecode.py
@@ -17,6 +17,6 @@ def read_and_decode(filename):
 
     img = tf.decode_raw(features['data_raw'], tf.float64)
     img = tf.reshape(img, [8, 550, 2])
-    label = tf.cast(features['label'], tf.int32)
+    label = tf.cast(features['label'], tf.int64)
 
     return img, label
